+++ wms-common/src/main/java/com/mz/common/service/IService.java
@@ -75,7 +75,7 @@ public interface IService<T> {
      * @param id
      * @return
      */
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
 
     /**
      * 更新（自定义参数）
